+++ include/kernel/types.h
@@ -149,8 +149,16 @@ typedef enum {
 #define ENOTEMPTY      39       /* Directory not empty */
 #define ELOOP          40       /* Too many symbolic links */
 #define EOVERFLOW      75       /* Value too large for defined data type */
+#define EDESTADDRREQ   89       /* Destination address required */
+#define EMSGSIZE       90       /* Message too long */
+#define EPROTONOSUPPORT 93      /* Protocol not supported */
+#define EAFNOSUPPORT   97       /* Address family not supported */
 #define EADDRINUSE     98       /* Address already in use */
+#define EADDRNOTAVAIL  99       /* Cannot assign requested address */
 #define ENETDOWN      100       /* Network is down */
+#define ENETUNREACH   101       /* Network is unreachable */
+#define ECONNABORTED  103       /* Software caused connection abort */
+#define ECONNRESET    104       /* Connection reset by peer */
 #define ENOTCONN      107       /* Transport endpoint is not connected */
 #define ETIMEDOUT     110       /* Connection timed out */
 #define EINPROGRESS   115       /* Operation now in progress */